Разделить ответственность
