+++ MovieQuiz/Presentation/MovieQuizViewController.swift
@@ -9,103 +9,28 @@ final class MovieQuizViewController: UIViewController {
     @IBOutlet weak private var yesButton: UIButton!
     @IBOutlet weak private var noButton: UIButton!
     
-    // Структура для отображения результатов в алерте
-    private struct QuizResultsViewModel {
-        let title: String
-        let text: String
-        let buttonText: String
-        init(title: String, text: String, buttonText: String) {
-            self.title = title
-            self.text = text
-            self.buttonText = buttonText
-        }
-    }
-    
-    // Структура для вопроса
-    private struct QuizQuestion {
-        let image: String
-        let text: String
-        let correctAnswer: Bool
-        init(image: String, text: String, correctAnswer: Bool) {
-            self.image = image
-            self.text = text
-            self.correctAnswer = correctAnswer
-        }
-    }
-    
-    // Структура для отображения вопроса на экране
-    private struct QuizStepViewModel {
-        let image: UIImage
-        let question: String
-        let questionNumber: String
-        init(image: UIImage, question: String, questionNumber: String) {
-            self.image = image
-            self.question = question
-            self.questionNumber = questionNumber
-        }
-    }
-    
-    // Массив mok`ов
-    private let questions: [QuizQuestion] = [
-        QuizQuestion(
-            image: "The Godfather",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Dark Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Kill Bill",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Avengers",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Deadpool",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "The Green Knight",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: true),
-        QuizQuestion(
-            image: "Old",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "The Ice Age Adventures of Buck Wild",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Tesla",
-            text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false),
-        QuizQuestion(
-            image: "Vivarium", text: "Рейтинг этого фильма больше чем 6?",
-            correctAnswer: false)
-    ]
-    
     // Переменная-счетчик текущего вопроса
     private var currentQuestionIndex: Int = 0
     // Переменная-счетчик количества правильных ответов
     private var correctAnswers: Int = 0
     
-    // Приватный метод конвертации mok`а в view-модель
+    private let questionsAmount: Int = 10
+    private var questionFactory: QuestionFactoryProtocol /*QuestionFactory*/ = QuestionFactory()
+    private var currentQuestion: QuizQuestion?
+    
+    // Приватный метод конвертации mock`а в view-модель
     private func convert(model: QuizQuestion) -> QuizStepViewModel {
         let questionStep = QuizStepViewModel(
             image: UIImage(named: model.image) ?? UIImage(),
             question: model.text,
-            questionNumber: "\(currentQuestionIndex + 1)/\(questions.count)")
+            questionNumber: "\(currentQuestionIndex + 1)/\(questionsAmount.count)")
         return questionStep
     }
     
-    // Приватный метод показа mok`а на экране
+    // Приватный метод показа mock`а на экране
     private func show(quiz step: QuizStepViewModel) {
         
-        // Задание значений элементам экран из view-модели mok`а
+        // Задание значений элементам экран из view-модели mock`а
         counterLabel.text = step.questionNumber
         imageView.image = step.image
         textLabel.text = step.question
@@ -143,10 +68,12 @@ final class MovieQuizViewController: UIViewController {
     private func showNextQuestionOrResults() {
         
         // Если текущий mok был последним
-        if currentQuestionIndex == questions.count - 1 {
+        if currentQuestionIndex == questionsAmount - 1 {
             
             // Текст в переменную для удобства редактирования
-            let text = "Ваш результат: \(String(correctAnswers))" + "/10"/* + "\n" + "Количество сыгранных квизов: 1" + "\n" + "Рекорд 0/0 (дата время)" + "\n" + "Средняя точность: 00.00%"*/
+            let text = correctAnswers == questionsAmount ? "Поздравляем, вы ответили на 10 из 10!" :
+            "Вы ответили на \(correctAnswers) из 10, попробуйте ещё раз!"
+            //            "Ваш результат: \(String(correctAnswers))" + "/10"/* + "\n" + "Количество сыгранных квизов: 1" + "\n" + "Рекорд 0/0 (дата время)" + "\n" + "Средняя точность: 00.00%"*/
             
             // Задание параметров модели алерта
             let viewModel = QuizResultsViewModel(
@@ -162,13 +89,20 @@ final class MovieQuizViewController: UIViewController {
             // Переход к следующему вопросу
             currentQuestionIndex += 1
             
-            // Константа для данных следующего mok`а
-            let nextQuestion = questions[currentQuestionIndex]
-            // Константа для модели mok`а
-            let viewModel = convert(model: nextQuestion)
+//            // Константа для данных следующего mock`а
+//            let nextQuestion = questions[currentQuestionIndex]
+//            // Константа для модели mok`а
+//            let viewModel = convert(model: nextQuestion)
+//            
+//            // Вызов метода для показа следующего mock`а на экране
+//            show(quiz: viewModel)
             
-            // Вызов метода для показа следующего mok`а на экране
-            show(quiz: viewModel)
+            if let nextQuestion = questionFactory.requestNextQuestion() {
+                currentQuestion = nextQuestion
+                let viewModel = convert(model: nextQuestion)
+
+                show(quiz: viewModel)
+            }
             
         }
     }
@@ -193,13 +127,21 @@ final class MovieQuizViewController: UIViewController {
             self.currentQuestionIndex = 0
             self.correctAnswers = 0
             
-            // Константа для выбора 0-го элемента из массива mok`ов
-            let firstQuestion = self.questions[self.currentQuestionIndex]
-            // Константа для модели mok`а первого вопроса
-            let viewModel = self.convert(model: firstQuestion)
+//            // Константа для выбора 0-го элемента из массива mok`ов
+//            let firstQuestion = self.questions[self.currentQuestionIndex]
+//            // Константа для модели mok`а первого вопроса
+//            let viewModel = self.convert(model: firstQuestion)
+//            
+//            // Вызов метода показа первого mok`а
+//            self.show(quiz: viewModel)
             
-            // Вызов метода показа первого mok`а
-            self.show(quiz: viewModel)
+            if let firstQuestion = self.questionFactory.requestNextQuestion() {
+                self.currentQuestion = firstQuestion
+                let viewModel = self.convert(model: firstQuestion)
+
+                self.show(quiz: viewModel)
+            }
+            
         }
         
         // Добавление действия к алерту
@@ -213,15 +155,27 @@ final class MovieQuizViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
-        // Показ первого вопроса
-        show(quiz: convert(model: questions[currentQuestionIndex]))
+//        // Показ первого вопроса
+//        show(quiz: convert(model: questions[currentQuestionIndex]))
+        
+        if let firstQuestion = questionFactory.requestNextQuestion() {
+            currentQuestion = firstQuestion
+            let viewModel = convert(model: firstQuestion)
+            show(quiz: viewModel)
+        }
+        
     }
     
     // Приватный метод выполняемый при нажатии кнопки ДА
     @IBAction private func yesButtonClicked(_ sender: UIButton) {
         
-        // Константа для хранения данных из текущего mok`а
-        let currentQuestion = questions[currentQuestionIndex]
+//        // Константа для хранения данных из текущего mock`а
+//        let currentQuestion = questions[currentQuestionIndex]
+        
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
+        
         // Константа для записи значения ответа пользователя на вопрос
         let givenAnswer = true
         
@@ -233,8 +187,13 @@ final class MovieQuizViewController: UIViewController {
     // Приватный метод выполняемый при нажатии кнопки НЕТ
     @IBAction private func noButtonClicked(_ sender: UIButton) {
         
-        // Константа для хранения данных из текущего mok`а
-        let currentQuestion = questions[currentQuestionIndex]
+//        // Константа для хранения данных из текущего mock`а
+//        let currentQuestion = questions[currentQuestionIndex]
+        
+        guard let currentQuestion = currentQuestion else {
+            return
+        }
+        
         // Константа для записи значения ответа пользователя на вопрос
         let givenAnswer = false
         
@@ -243,67 +202,3 @@ final class MovieQuizViewController: UIViewController {
         
     }
 }
-
-/*
- Mock-данные
- 
- 
- Картинка: The Godfather
- Настоящий рейтинг: 9,2
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: The Dark Knight
- Настоящий рейтинг: 9
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: Kill Bill
- Настоящий рейтинг: 8,1
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: The Avengers
- Настоящий рейтинг: 8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: Deadpool
- Настоящий рейтинг: 8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: The Green Knight
- Настоящий рейтинг: 6,6
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: ДА
- 
- 
- Картинка: Old
- Настоящий рейтинг: 5,8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- 
- 
- Картинка: The Ice Age Adventures of Buck Wild
- Настоящий рейтинг: 4,3
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- 
- 
- Картинка: Tesla
- Настоящий рейтинг: 5,1
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- 
- 
- Картинка: Vivarium
- Настоящий рейтинг: 5,8
- Вопрос: Рейтинг этого фильма больше чем 6?
- Ответ: НЕТ
- */
